feat: add event provider
