+++ simaple/simulate/event.py
@@ -1,43 +1,46 @@
+from abc import ABCMeta, abstractmethod
+
 from simaple.simulate.base import Event
 
 
-def rejected_event(name: str):
-    return Event(name=name, tag="global.reject")
+class EventProvider(metaclass=ABCMeta):
+    @abstractmethod
+    def elapsed(self, time) -> Event:
+        ...
+
+    @abstractmethod
+    def rejected(self) -> Event:
+        ...
+
+    @abstractmethod
+    def delayed(self, time) -> Event:
+        ...
+
+    @abstractmethod
+    def dealt(self, damage: float, hit: float, delay: float) -> Event:
+        ...
 
 
-def elapsed_event(name: str, time: float):
-    return Event(name=name, tag="global.elapsed", payload={"time": time})
+class NamedEventProvider(EventProvider):
+    def __init__(self, name: str):
+        self._name = name
 
+    def elapsed(self, time) -> Event:
+        return Event(name=self._name, tag="global.elapsed", payload={"time": time})
 
-def action_delay_triggered(name: str, time: float):
-    return Event(name=name, tag="global.delay", payload={"time": time})
+    def rejected(self) -> Event:
+        return Event(name=self._name, tag="global.reject")
 
+    def delayed(self, time) -> Event:
+        return Event(name=self._name, tag="global.delay", payload={"time": time})
 
-def damage_event(name: str, damage: float, hit: float, delay: float):
-    return Event(
-        name=name,
-        tag="global.damage",
-        payload={
-            "damage": damage,
-            "hit": hit,
-            "delay": delay,
-        },
-    )
-
-
-def buff_used_event():
-    return Event(
-        tag="global.buff.used",
-    )
-
-
-def interval_damage_event(name: str, damage: float, hit: float, count: int):
-    return Event(
-        name=name,
-        tag="global.damage.interval",
-        payload={
-            "damage": damage,
-            "hit": hit,
-            "count": count,
-        },
-    )
+    def dealt(self, damage: float, hit: float, delay: float) -> Event:
+        return Event(
+            name=self._name,
+            tag="global.damage",
+            payload={
+                "damage": damage,
+                "hit": hit,
+                "delay": delay,
+            },
+        )
